split out terrier stemmer, tokeniser, and stopwords
