+++ pyterrier/terrier/__init__.py
@@ -3,6 +3,9 @@ from pyterrier.terrier import java
 from pyterrier.terrier.java import configure, set_version, set_helper_version, enable_prf, extend_classpath, J, set_property, set_properties, run, version, check_version, check_helper_version
 from pyterrier.terrier.retriever import BatchRetrieve, FeaturesBatchRetrieve
 from pyterrier.terrier.index_factory import IndexFactory
+from pyterrier.terrier.stemmer import TerrierStemmer
+from pyterrier.terrier.tokeniser import TerrierTokeniser
+from pyterrier.terrier.stopwords import TerrierStopwords
 from pyterrier.terrier import index
 from pyterrier.terrier.index import TerrierIndexer, FilesIndexer, TRECCollectionIndexer, DFIndexer, DFIndexUtils, IterDictIndexer, IndexingType, TerrierStemmer, TerrierStopwords, TerrierTokeniser
 from pyterrier.terrier import rewrite
@@ -18,11 +21,12 @@ __all__ = [
     'BatchRetrieve', 'TerrierRetrieve', 'FeaturesBatchRetrieve',
 
     # indexing
-    'index', 'TerrierIndexer', 'FilesIndexer', 'TRECCollectionIndexer', 'DFIndexer', 'DFIndexUtils', 'IterDictIndexer', 'IndexingType', 'TerrierStemmer', 'TerrierStopwords', 'TerrierTokeniser',
+    'index', 'TerrierIndexer', 'FilesIndexer', 'TRECCollectionIndexer', 'DFIndexer', 'DFIndexUtils', 'IterDictIndexer', 'IndexingType',
 
     # rewriting
     'rewrite',
 
     # misc
+    'TerrierStemmer', 'TerrierStopwords', 'TerrierTokeniser',
     'IndexFactory', 'set_property', 'set_properties', 'run', 
 ]
